Extract common test data class
