+++ extensions/identity-hub/src/test/java/org/eclipse/dataspaceconnector/identityhub/processor/CollectionsWriteProcessorTest.java
@@ -16,6 +16,7 @@ package org.eclipse.dataspaceconnector.identityhub.processor;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.javafaker.Faker;
+import org.eclipse.dataspaceconnector.identityhub.TestData;
 import org.eclipse.dataspaceconnector.identityhub.dtos.MessageResponseObject;
 import org.eclipse.dataspaceconnector.identityhub.dtos.MessageStatus;
 import org.eclipse.dataspaceconnector.identityhub.dtos.credentials.VerifiableCredential;
@@ -49,18 +50,17 @@ public class CollectionsWriteProcessorTest {
     @Test
     void writeCredentials() throws Exception {
         // Arrange
-        var credentialId = FAKER.internet().uuid();
-        var verifiableCredentialMap = Map.of("id", credentialId);
-        var data = OBJECT_MAPPER.writeValueAsString(verifiableCredentialMap).getBytes(StandardCharsets.UTF_8);
+        var verifiableCredential = TestData.createVerifiableCredential();
+        var data = OBJECT_MAPPER.writeValueAsString(verifiableCredential).getBytes(StandardCharsets.UTF_8);
 
         // Act
         var result = writeProcessor.process(data);
 
         // Assert
         var expectedResult = MessageResponseObject.Builder.newInstance().messageId(MESSAGE_ID_VALUE).status(MessageStatus.OK).build();
-        var expectedVerifiableCredential = VerifiableCredential.Builder.newInstance().id(credentialId).build();
+
         assertThat(result).usingRecursiveComparison().isEqualTo(expectedResult);
-        assertThat(identityHubStore.getAll()).usingRecursiveFieldByFieldElementComparator().containsExactly(expectedVerifiableCredential);
+        assertThat(identityHubStore.getAll()).usingRecursiveFieldByFieldElementComparator().containsExactly(verifiableCredential);
     }
 
     @Test
